RFC: Initial sketch up of output driver
This is an initial sketch up for an output driver API for RFC.
Parties interested in using an output transport besides the provided
I2C should comment.

NOTES:

1. THIS HAS NOT BEEN TESTED AT ALL. IT HAS NOT BEEN COMPILED TESTED.

2. This starts the process of reducing copies on the output side. The
change assembles the HIF packets as soon as data is generated from the
algs. Prior code converts the alg output into an intermediate format which
in turn is sent via the RTOS so it can be converted to the HIF packet.

3. This commit adds the data output path. In particular, the data
input (control packet) path is not fully fleshed out. It needs to be
factored out of the i2c_slavecomm task into a general purpose input
path.

4. The I2C slavecomm task is not really needed after this change. It is
kept in this sketch and (and will be in the initial implementation) for
debugging reasons.

5. Driver_OUT.h contains a description of the API. It should be possible
to handle the output without the driver knowning anything about the HIF
packets.

Signed-off-by: Hunyue Yau <hy-git@hy-research.com>